requirements.txt
fastapi
keras
numpy
Pillow
tensorflow
tensorflow_intel
uvicorn
python-multipart
